requirements.txt
absl-py==1.4.0
asgiref==3.7.2
astunparse==1.6.3
backports.zoneinfo==0.2.1
bayesian-optimization==1.4.2
blobfile==2.0.2
cachetools==5.3.0
certifi==2023.7.22
charset-normalizer==3.2.0
cma==3.3.0
cmake==3.27.0
colorama==0.4.6
contourpy==1.1.0
cycler==0.11.0
Django==4.2.3
efficientnet-pytorch==0.7.1
filelock==3.12.2
fonttools==4.42.1
gast==0.4.0
google-auth==2.17.1
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.53.0
h5py==3.8.0
idna==3.4
image==1.5.33
importlib-metadata==6.1.0
importlib-resources==6.0.1
jax==0.4.8
Jinja2==3.1.2
jsonpatch==1.32
jsonpointer==2.3
Keras-Preprocessing==1.1.2
kiwisolver==1.4.5
libclang==16.0.0
lit==16.0.6
lxml==4.9.3
Markdown==3.4.3
MarkupSafe==2.1.2
matplotlib==3.7.2
ml-dtypes==0.0.4
mpi4py==3.1.4
mpmath==1.3.0
networkx==3.1
nevergrad==0.6.0
numpy==1.23.5
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.2
opencv-python==4.8.0.76
opt-einsum==3.3.0
packaging==23.1
pandas==2.0.0
Pillow==10.0.0
pip==23.2.1
protobuf==4.22.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycryptodomex==3.18.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytorch-fid==0.3.0
pytorch-fid-wrapper==0.0.4
pytz==2023.3
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scipy==1.10.1
setuptools==68.0.0
six==1.16.0
sqlparse==0.4.4
sympy==1.12
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
tensorflow-gpu==2.7.0
tensorflow-io-gcs-filesystem==0.32.0
termcolor==2.2.0
torch==2.0.1
torchsummary==1.5.1
torchvision==0.15.2
tornado==6.2
tqdm==4.65.0
triton==2.0.0
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
visdom==0.2.4
websocket-client==1.5.1
Werkzeug==2.2.3
wheel==0.38.4
wrapt==1.14.1
zipp==3.16.2
